feat: add ngrok server functionality
This commit introduces the `ngrok` server functionality to the main application. The `ngrok` package was imported and used to set up a public URL for the local server. Also, a new server running function was created to handle the ngrok process and server shutdown. Lastly, `pyngrok` was added to the project requirements.