feat: Update paths and mount points for static files
Updated the paths and mount points for static files in the .gitignore and main.py files. Also, adjusted the model directory path in main.py and the model loading path in model.js. Additionally, updated the working directory and script name in the run configuration. Lastly, modified the script source path in the index.html and added new libraries in vixevia.iml.